Fix the organization deletion method by adding validation

+++ src/main/java/com/example/backend/service/OrganizationService.java
@@ -4,6 +4,7 @@ import com.example.backend.model.Organization;
 import com.example.backend.repository.OrganizationRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+
 import java.util.HashSet;
 import java.util.Optional;
 import java.util.Set;
@@ -21,7 +22,7 @@ public class OrganizationService {
     public Optional<Organization> getOrganizationById(Long id) {
         if (organizationRepository.existsById(id)) {
             return organizationRepository.findById(id);
-        }else{
+        } else {
             throw new IllegalArgumentException("The Organization with inserted id doesn't exist");
         }
     }
@@ -32,7 +33,11 @@ public class OrganizationService {
         } else throw new IllegalArgumentException("The Organization name or id already exists!");
     }
 
-    public void removeOrganization(Long id) {
-        organizationRepository.deleteById(id);
+    public void removeOrganizationById(Long id) {
+        if (organizationRepository.existsById(id)) {
+            organizationRepository.deleteById(id);
+        } else {
+            throw new IllegalArgumentException("The Organization with inserted id doesn't exist");
+        }
     }
 }
